fix(5-1): rename data folder to offers in store
