+++ src/store/offers/selector.js
